requirements.txt
fabric
invoke
jinja2
PyNaCl
fastapi
PyMySQL
urllib3
uvicorn
pydantic
pyroute2
requests
sqlmodel
tenacity
SQLAlchemy
python-jose
apache-libcloud
wireguard-tools
google-cloud-secret-manager
cloud-sql-python-connector[pymysql]
